+++ readium/shared/r2-shared/src/test/java/org/readium/r2/shared/util/mediatype/MediaTypeTest.kt
@@ -1,16 +1,7 @@
-/*
- * Module: r2-shared-kotlin
- * Developers: Mickaël Menu
- *
- * Copyright (c) 2020. Readium Foundation. All rights reserved.
- * Use of this source code is governed by a BSD-style license which is detailed in the
- * LICENSE file present in the project repository where this source code is maintained.
- */
+package org.readium.r2.shared.util.mediatype
 
-package org.readium.r2.shared.format
-
-import org.junit.Assert.*
 import org.junit.Test
+import kotlin.test.*
 
 class MediaTypeTest {