+++ readium/shared/CHANGELOG.md
@@ -9,8 +9,11 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 * Upgraded to Kotlin 1.4.10.
-* `ContentLayout` is deprecated, use `publication.metadata.effectiveReadingProgression` to determine the reading
-progression of a publication instead.
+* `Format` got merged into `MediaType`, to simplify the media type APIs.
+  * You can use `MediaType.of()` to sniff the type of a file or bytes.
+  * `MediaType` has now optional `name` and `fileExtension` properties.
+  * Some publication formats can be represented by several media type aliases. Using `mediaType.canonicalized` will give you the canonical media type to use, for example when persisting the file type in a database. All Readium APIs are already returning canonical media types, so it only matters if you create a `MediaType` yourself from its string representation.
+* `ContentLayout` is deprecated, use `publication.metadata.effectiveReadingProgression` to determine the reading progression of a publication instead.
 
 
 ## [2.0.0-alpha.2]
@@ -47,7 +50,7 @@ progression of a publication instead.
 * The new [Format API](https://github.com/readium/architecture/blob/master/proposals/001-format-api.md) simplifies the detection of file formats, including known publication formats such as EPUB and PDF.
   * [A format can be "sniffed"](https://github.com/readium/architecture/blob/master/proposals/001-format-api.md#sniffing-the-format-of-raw-bytes) from files, raw bytes or even HTTP responses.
   * Reading apps are welcome to [extend the API with custom formats](https://github.com/readium/architecture/blob/master/proposals/001-format-api.md#supporting-a-custom-format).
-  * Using `Link.mediaType?.matches()` is now recommended [to safely check the type of a resource](https://github.com/readium/architecture/blob/master/proposals/001-format-api.md#mediatype-class).
+  * Using `Link.mediaType.matches()` is now recommended [to safely check the type of a resource](https://github.com/readium/architecture/blob/master/proposals/001-format-api.md#mediatype-class).
   * [More details about the Kotlin implementation can be found in the pull request.](https://github.com/readium/r2-shared-kotlin/pull/100)
 * In `Publication` shared models:
   * Support for the [Presentation Hints](https://readium.org/webpub-manifest/extensions/presentation.html) extension.